Consolidate box sampling and target materialization
